+++ src/ambition/version_details.cc
@@ -26,6 +26,7 @@ std::vector<VersionUpdates> versionDetails = {
       {
         .heading = _(HEADING_FIXED),
         .items = {
+          _("Fixed builders sometimes going to their buildings' rally points instead of to the building where they were requested by the request builder button.  The workaround for this means that builders will always ignore rally points, for the time being."),
           _("Fixed non-English tutorial descriptions getting cut off at the bottom, by adding a scroll bar and buttons, so that all the text can be scrolled and read."),
         },
       },