+++ include/ambition/7kaaInterface/control.hh
@@ -59,6 +59,15 @@ void pasteFromClipboard(
   const unsigned int maximumSize
 );
 
+bool preventReplayDeleteStopDesync(
+  const UnitCaravan* _7kaaCaravan,
+  const int _7kaaStopId
+);
+bool preventReplayDeleteStopDesync(
+  const UnitMarine* _7kaaTradeShip,
+  const int _7kaaStopId
+);
+
 bool preventReplaySetStopPickupDesync(
   const UnitCaravan* _7kaaCaravan,
   const int _7kaaStopId